FIR: keep `@EnhancedNullability` when substituting
There was a special case where it was kept on definitely not null types,
but this commit extends that to all substitutions.

I'm not sure if this actually matters since the backend seems to do a
pretty good job of constructing bridges regardless. The one new
IMPLICIT_NOTNULL type operator looks like it might be observable.

This behavior is more FE1.0-like in any case.