FE: enhance arguments even if they don't match the parameter list
e.g. B in C<A, B> given class C<T>

This is mostly for some tests that use FIR and package-private classes
in differently named files, as the test framework passes the java files
as a directory and FIR assumes each class must be in an equivalently
named file, so the class cannot be loaded. Oops. TODO: fix that maybe?