FIR: use other type arguments when enhancing unbounded `?`
`?` is replaced with `? extends U` where `U` is the type parameter's
upper bound, and it may depend on other type parameters, so the correct
transformation needs to use them. FE1.0 does not do that in general (not
just in enhancement), leading to weird errors:

  public interface J<T, V extends T> {
    V foo();
  }

  fun f(j: J<String, *>): String? = j.foo() // got Any!, not String!

FIR successfully compiles this code, so enhancement should probably not
break it either.