[Gradle] [MPP] Fix 'getDependenciesTransformation' for ^KT-46273
This setup requires a 'jvm like' target and an orphan source set, which
depends on 'commonTest'. This orphan will not participate in any
compilation and therefore will not have any platform.

The code expected at least one test task for this source set
and therefore at least one test framework.

^KT-46273 Verification Pending