FIR checker: accept unambiguous `AMBIGUOUS_SUPER`
Before this change, FIR and FE1.0 both rejects the following code
by reporting `super` as ambiguous and requires user to explicitly
annotate `super` with `super<B>`.

interface A {
  fun foo() {}
}
interface B : A
interface C : B, A {
  fun test() {
    super.foo()
  }
}

However, this seems to be redundant because the alternative `super<A>`
is not valid due to QUALIFIED_SUPERTYPE_EXTENDED_BY_OTHER_SUPERTYPE.
Hence this change simply allows such unannotated `super`.