Move system property initialization near application initialization
Previously there were several random places where we do initialization
of the System properties that later were used in the IDEA core that
we are using in the compiler. Moved the initialization of such
properties right before the Application initialization.
Also add assertion that there wer no Application initialized before,
because it cannot be correct behaviour -- a lot of IDEA staff is
initialized in the static field and having two running application
in the same process is error.

P.s. see previous commit to see beginning of the story.
Previous commit was leaved "as is" so we could pick it into release
branches easily