Partially revert previous commit
We need to get rid of hard reference to the LanguageLevel entries
because we need 213 plugin for IDEA, but let's minimize the change
and pick JDK_15 in 1.6.0 branch as before.
P.s. it seems like for old IDEA version there is a difference between
JDK_15_PREVIEW and JDK_15 so don't change it for 1.6.0 branch