Model mapping: ignore attempts to add a compilation's default source set
The variant-mapped compilation details don't store "explicitly added"
source sets; instead, this class redirects the calls
to `source(...)` to the default source set's `dependsOn`. It could be
potentially valid to call `source(defaultSourceSet)`, as
it should be idempotent, but this implementation would throw an error
due to a detected dependsOn cycle (foo -> foo), so it needs to filter
out calls to `source(defaultSourceSet)`.