Native build: make distCompiler no-op when using custom dist
Custom dist is already in place and has the compiler, so we don't have
to build and copy the compiler to dist.

Moreover, if we do copy it, it might overwrite the compiler files
already loaded by this Gradle process (including the jar loaded to the
custom classloader), causing hard-to-debug errors.

To workaround this, we disable the distCompiler task and don't add any
dependencies to it when custom dist is used.