Remove compose plugin frontend dependencies in backend
Remove calls to `FunctionDescriptor.allowsComposableCalls` in the
compose IR plugin. These calls implicitly depend on the frontend
analysis through binding context slices which will not be
available with the FIR frontend.

The main change in this CL is to enable IrInlineReferenceLocator to
detect lambdas capable of composer capture. It also fixes a few minor
issues along the way:

- Detect inline lambda arguments to the two argument Array constructor
  calls. (This still needs support in the frontend part of the Compose
  plugin!)
- Don't transform the block arguments to remember in
  ComposableFunctionBodyTransformer, since they are marked with
  @DisallowComposableCalls
- Mark the block argument to Composer.cache with
  @DisallowComposableCalls, since such calls occur after the remember
  intrinsic transformation and could otherwise confuse later passes

Apart from that, the frontend analysis information can be incomplete,
since it is generated on demand in response to composable calls in
ComposableCallChecker. This can lead to some missed optimizations in the
backend. This is fixed by recomputing the relevant information in the
backend.

Bug: 229716591 ( https://issuetracker.google.com/issues/229716591 )
Fixes: 229719233 ( https://issuetracker.google.com/issues/229719233 ), 229719460
Test: ./gradlew :compose:c:c-h:i-t:tDUT
Change-Id: Iaa7a5cfffbf6cdcc23c4e3cd90adee649c627d86 ( https://android-review.git.corp.google.com/q/Iaa7a5cfffbf6cdcc23c4e3cd90adee649c627d86 )