Add support for using composable target annotation at file scope
`@ComposableTarget` and annotations marked `@ComposableTargetMarker` are
now looked for as `@file:` annotations. Adding a composable target
annotation at the file scope declares that all `2Compsoable` functions
in this file are, by default, assumed to be targetting the applier implied
by the annotations. For examplea, using `@file:UiComposable` will cause
all `@Composable` functions in the file to be assumed to be targetting
the Compose UI applier.

Relnote: """The `@ComposableTarget` annotation and annotations marked by
`@ComposableTargetMarker` can now be used at the file scope using the
`@file` prefix. Using a target annotation at the file scope will cause the
compiler to assume all composable functions in the file are intended to be
targeth the associated applier. For example, using `@file:UiComposable`
declares that all `@Composable` functions target the Compose UI
applier. A function that needs to target another applier must explicity
supply the target marker annotation for the desired applier.
"""

Fixes: 228078200 ( https://issuetracker.google.com/issues/228078200 )
Test: ./gradlew :compose:c:c-h:i-t:tDUT
Change-Id: I40804e71dcc931b788c101a0be90ae1d6fc7eba1 ( https://android-review.git.corp.google.com/q/I40804e71dcc931b788c101a0be90ae1d6fc7eba1 )