KT-52669: Fix fallback logic in IncrementalCompilerRunner
The current logic works as follows:
   - Try either incremental compilation or non-incremental compilation
   - If the above (or any of its surrounding work) fails, fall back to
     non-incremental compilation

This means we may perform non-incremental compilation twice.

This commit will fix that logic so that we fall back to non-incremental
compilation only if *incremental compilation* fails.

A nice consequence of this change is that it also resolves the critical
bugs described at KT-52669 (which occur because the current logic is
flawed).

^KT-52669: In Progress