Fix group generation for loops containing conditional code
In some cases, when the body of a loop contained conditional code
a group not inserted where it was required around conditional
statements in the body of the loop.

This change enforces that the conditional code is surrounded by
a group if the containing loop group can be elided.

Fixes: 205590513 ( https://issuetracker.google.com/issues/205590513 )
Test:  ./gradlew :compose:compiler:c-h:i-t:tDUT :compose:r:r:i-t:cC

Change-Id: I5f384fe5a61bc924bf973b2cf3835cb2535bea28 ( https://android-review.googlesource.com/q/I5f384fe5a61bc924bf973b2cf3835cb2535bea28 )

Moved from: https://github.com/androidx/androidx/commit/3d4ed676cb9a6f9ac352788d92ff726f790af6aa