Fix generating closing groups for non-local returns
If an inline lambda contained a non-local return then the code generated
for the lambda incorrectly closed the groups of the caller if the caller
was also a composable function.

As it cannot be known at the point of the return how many groups need to
be closed, this change adds a `endToMarker` and `currentMarker` calls to
the composer that closes the groups up to, but not including, the current
parent group for which the marker was obtained.

Calls to the new methods require the module be comiled with a reference to
a runtime that contains the new methods.

Relnote: N/A
Fixes: 209619743 ( https://issuetracker.google.com/issues/209619743 )
Test: ./gradlew :compose:compiler:c-h:i-t:tDUT
Change-Id: I68832dcc1164826a359c6a7494d7aa8d398895fb ( https://android-review.googlesource.com/q/I68832dcc1164826a359c6a7494d7aa8d398895fb )

Moved from: https://github.com/androidx/androidx/commit/b6f590c5a33daf9ec90470bd9f6b51eac88f1eef